Repository added

+++ GroceryShop.Entity/Category.cs
@@ -10,7 +10,7 @@ namespace GroceryShop.Entity
     {
         public string AppId { get; set; }
         public string Name { get; set; }
-        public float Discount { get; set; }
+        public double Discount { get; set; }
         public string Description { get; set; }
         public string UpdatedAt(string datetime = "GETDATE()")
         {
